chore: address SonarCloud findings from PR review
Two missed S1481 cascades from the original sweep: deleting consumers
left the source locals (PostgresQueryBuilder.foreignKey,
MockDataService.persType) orphaned. Drop both.

DRY the ValidateWriteOpInput / GetFromAndToEntities pair that was
inlined into ServiceOwnerConnectionService in the #3108 fix.
Extract to ConnectionWriteValidation internal static helpers under
Altinn.AccessMgmt.Core.Validation; both ConnectionService (one-line
forwarders to keep existing call sites stable) and
ServiceOwnerConnectionService now share the implementation. Silences
SonarCloud's new_duplicated_lines_density gate; net -16 lines.

Skipping the 4× S112 findings: AssignmentService.cs:398 matches the
established throw-new-Exception style in the surrounding method, and
DelegationService.cs:192/291/400 are pre-existing throw expressions
that Sonar diff-flagged because S1481 changed `var x =` to `_ =` on
the same line. Both classes belong in a future S112 sweep that
treats the 27+ sites consistently, not in this PR.
